Time: 4 ms (27.82%), Space: 111.3 MB (40.61%) - LeetHub

+++ 2432-number-of-zero-filled-subarrays/2432-number-of-zero-filled-subarrays.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
      long long ans = 0;
-     long long n = nums.size();
-     long long l = 0;
-      long long r = 0;
-     for(long long i = 0; i<n; i++){
+     int n = nums.size();
+     int l = 0;
+      int r = 0;
+     for(int i = 0; i<n; i++){
         if(nums[i] == 0){
             long long len = 0;
             while(i<n &&nums[i] == 0) {
